@@ -52,8 +54,13 @@ pytest-cov = "^4.1"
 "Bug Tracker" = "https://github.com/tare/GPMicrobiome/issues"
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py310"
 fix = true
+exclude = ["LuxRep/"]
+force-exclude = true
+
+[tool.ruff.lint]
+preview = true
 fixable = [
     "I001",  # https://docs.astral.sh/ruff/rules/unsorted-imports/
 ]
