@@ -1,5 +1,6 @@
 """__init__.py."""
+
 from gpmicrobiome.inference import run_nuts
 from gpmicrobiome.utils import get_mcmc_summary
 
-__all__ = ["run_nuts", "get_mcmc_summary"]
+__all__ = ["get_mcmc_summary", "run_nuts"]
